requirements.txt
streamlit==1.32.0
Pillow==10.0.0
numpy==1.24.3
pytest==7.4.0
python-dotenv==1.0.0
openai==1.12.0